+++ tools/src/main/java/it/eng/tools/model/DSpaceConstants.java
@@ -15,7 +15,7 @@ public interface DSpaceConstants {
 	}
 	
 	public static enum Operators {
-		EQ, GT, GTEQ, HAS_PARENT, IS_A, IS_ALL_OF, IS_ANY_OF, IS_NONE_OF, IS_PART_OF, LT, TERM_LTEQ, NEQ;
+		EQ, GT, GTEQ, HAS_PARENT, IS_A, IS_ALL_OF, IS_ANY_OF, IS_NONE_OF, IS_PART_OF, LT, LTEQ, NEQ;
 	}
 
 	public static final String DSPACE = "dspace:";